# Validate HTML in CI