--- a/package.json
+++ b/package.json
@@ -69,6 +69,7 @@
     "stylelint": "10.0.1",
     "svgo": "1.2.2",
     "swagger-parser": "6.0.5",
+    "vnu-jar": "^19.6.7",
     "webpack-dev-server": "3.5.1"
   },
   "scripts": {