--- a/tools/circleci/images.yml
+++ b/tools/circleci/images.yml
@@ -1,7 +1,3 @@
-trusty:
-  base_image: buildpack-deps:trusty-scm
-  extra_packages: python-virtualenv postgresql-9.3
-
 xenial:
   base_image: buildpack-deps:xenial-scm
   extra_packages: virtualenv postgresql-9.5