--- a/tools/test-help-documentation
+++ b/tools/test-help-documentation
@@ -1,9 +1,10 @@
 #!/usr/bin/env python3
 import argparse
+import contextlib
 import os
 import sys
 import subprocess
-from typing import List
+from typing import Iterator
 
 ZULIP_PATH = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
 
@@ -30,13 +31,28 @@ os.makedirs('var/help-documentation', exist_ok=True)
 LOG_FILE = 'var/help-documentation/server.log'
 external_host = "localhost:9981"
 
-extra_args = []  # type: List[str]
+extra_args = ['-a', 'validate_html=set']
 
 if options.skip_external_link_check:
-    extra_args = ['-a', 'skip_external=set']
+    extra_args += ['-a', 'skip_external=set']
 
-with test_server_running(options.force, external_host, log_file=LOG_FILE,
-                         dots=True, use_db=True):
+@contextlib.contextmanager
+def vnu_servlet() -> Iterator[None]:
+    with subprocess.Popen([
+        'java', '-cp',
+        os.path.join(
+            os.path.dirname(__file__),
+            '../node_modules/vnu-jar/build/dist/vnu.jar',
+        ),
+        'nu.validator.servlet.Main',
+        '9988',
+    ]) as proc:
+        yield
+        proc.terminate()
+
+with vnu_servlet(), \
+    test_server_running(options.force, external_host, log_file=LOG_FILE,
+                        dots=True, use_db=True):
     ret_help_doc = subprocess.call(['scrapy', 'crawl_with_status'] + extra_args +
                                    ['help_documentation_crawler'],
                                    cwd='tools/documentation_crawler')