The first three commits resolve many but not all of the outstanding HTML validation problems. They can be merged at any time.

The last two commits add HTML validation to `test-documentation` and `test-help-documentation`, but they’ll have to wait until we kill Trusty because `vnu.jar` requires Java 8. Many real HTML errors remain whitelisted for now, but at least we can start making sure we don’t get new ones.

**Testing Plan:** Run `tools/test-documentation --skip-check-links`, `tools/test-help-documentation --skip-external-links`.