@timabbott 2019-04-23
I merged a couple commits.  For `FinalizeOpenGraphDescription`, I'm not sure that's the right fix -- how will open graph tags with the escaped content appear?  (Also, roughly how many /help/ articles are affected?).

We might want to make changes to our tests and/or changes to the /help/ articles themselves if this makes articles not display correctly.