@andersk 2019-04-23
HTML attributes _always_ need to be HTML-escaped. There are no exceptions—not even for cases you might think to be weird like a `<button onclick=…>` attribute, and certainly not for a bog-standard `<meta content=…>` attribute. I don’t understand the concern about whether the escaped attributes will appear correctly. Without escaping, they literally aren’t even attributes.

Affected pages:

[/api/incoming-webhooks-walkthrough](https://chat.zulip.org/api/incoming-webhooks-walkthrough) ([validate](https://validator.nu/?doc=https%3A%2F%2Fchat.zulip.org%2Fapi%2Fincoming-webhooks-walkthrough))
[/api/register-queue](https://chat.zulip.org/api/register-queue) ([validate](https://validator.nu/?doc=https%3A%2F%2Fchat.zulip.org%2Fapi%2Fregister-queue))
[/api/update-message](https://chat.zulip.org/api/update-message) ([validate](https://validator.nu/?doc=https%3A%2F%2Fchat.zulip.org%2Fapi%2Fupdate-message))
[/help/change-your-name](https://chat.zulip.org/help/change-your-name) ([validate](https://validator.nu/?doc=https%3A%2F%2Fchat.zulip.org%2Fhelp%2Fchange-your-name))
[/help/desktop-app-install-guide](https://chat.zulip.org/help/desktop-app-install-guide) ([validate](https://validator.nu/?doc=https%3A%2F%2Fchat.zulip.org%2Fhelp%2Fdesktop-app-install-guide))
[/help/night-mode](https://chat.zulip.org/help/night-mode) ([validate](https://validator.nu/?doc=https%3A%2F%2Fchat.zulip.org%2Fhelp%2Fnight-mode))
[/help/status-and-availability](https://chat.zulip.org/help/status-and-availability) ([validate](https://validator.nu/?doc=https%3A%2F%2Fchat.zulip.org%2Fhelp%2Fstatus-and-availability))