@andersk 2019-04-23
By the way, it turns out all our OpenGraph tags are [busted anyway](https://developers.facebook.com/tools/debug/sharing/?q=https%3A%2F%2Fchat.zulip.org%2Fapi%2Fincoming-webhooks-walkthrough) because we always set `og:url` to point to the home page (which redirects to /login, whose `og:url` points back to the home page).