@timabbott 2019-04-23
Oh, I wasn't thinking -- this was just generic worry about double-escaping.  Makes sense, merged the open graph commit.  

I guess two embed preview implementations we used when testing the /help/ open graph stuff didn't use `og:url`, sigh.

Let's open a separate issue for `og:url` -- I think we either want to delete it (if that DTRT) or specify the current page's address (I think relative links don't work?); I'm not sure we have a use case for pointing to a different canonical URL from the current one right now.