@timabbott 2019-04-23
For the remaining two commits, we'll need to (1) fix the CI failure and (2) I'd like to have data on runtime and thus CI perf impact since that impacts where we'll want to run it.