@andersk 2019-04-23
CI failure: Yeah, see the comment above about Java 8 on Trusty. We could pull in Java 8 from a PPA on Trusty, or arrange for validation to be skipped on Trusty, or wait until we kill Trusty.

Timing: on my laptop, `tools/test-documentation --skip-check-links` goes from 9 seconds to 15 seconds, and `tools/test-help-documentation --skip-external-links` goes from 40 seconds to 48 seconds.