@timabbott 2019-04-23
OK, cool, sounds like timing isn't a material issue.

For CI, I think we should do the project of removing trusty support from master.  There's a bunch of things that are going to make that unpleasant over the next few weeks (e.g. Python dependencies have started dropping Python 3.4 support as that reached EOL too).  

I think the first and most annoying step there is to upgrade the Linux/Vagrant development environment to run a newer Ubuntu, which may be somewhat annoying because I think we might need to replace `vagrant-lxc` with their `libvirt` integration (not sure, but I am not sure there's a bionic prebuilt guest box for the LXC provider).  I think 90% of the work here is "Get a Vagrant Bionic guest running on Linux", so could be a good project for you as it doesn't require historical knowledge of how we set things up.