@timabbott 2019-05-29
I guess we can rebase this on the "remove trusty from CI" branch.  

Though actually, now that I'm looking at this, I think it might be a good idea regardless to move the external links tests in `tools/ci/backend` to only run in a single suite (maybe `xenial`), rather than all 3, since it's totally a waste of resources to be doing that link testing on every platform (similar to frontend tests).

Finally, for validating the main webapp's HTML, I had the thought that the `test-help-documentation` system is already booting a test webserver, and we could probably figure out how to make the system at least load the main webapp HTML (potentially using a "public archive" code path that doesn't involve dealing with authentication, though in theory just clicking a login link in the dev environment).   Would only test backend templates without JS, but that's something.