@andersk 2019-05-29
Rebased, although in addition to the validation regression since I first wrote this, we seem to be ignoring some pretty blatant errors:

```
Testing only internal links in documentation...
2019-05-29 23:49:04 [scrapy.core.scraper] ERROR: Spider error processing <GET file:///home/circleci/zulip/docs/_build/html/index.html> (referer: None)
Traceback (most recent call last):
  File "/srv/zulip-venv-cache/1dfe4580e212668e0395ceb9853ebc6f33e66f47/zulip-py3-venv/lib/python3.6/site-packages/scrapy/utils/defer.py", line 102, in iter_errback
    yield next(it)
  File "/srv/zulip-venv-cache/1dfe4580e212668e0395ceb9853ebc6f33e66f47/zulip-py3-venv/lib/python3.6/site-packages/scrapy/spidermiddlewares/offsite.py", line 29, in process_spider_output
    for x in result:
  File "/srv/zulip-venv-cache/1dfe4580e212668e0395ceb9853ebc6f33e66f47/zulip-py3-venv/lib/python3.6/site-packages/scrapy/spidermiddlewares/referer.py", line 339, in <genexpr>
    return (_set_referer(r) for r in result or ())
  File "/srv/zulip-venv-cache/1dfe4580e212668e0395ceb9853ebc6f33e66f47/zulip-py3-venv/lib/python3.6/site-packages/scrapy/spidermiddlewares/urllength.py", line 37, in <genexpr>
    return (r for r in result or () if _filter(r))
  File "/srv/zulip-venv-cache/1dfe4580e212668e0395ceb9853ebc6f33e66f47/zulip-py3-venv/lib/python3.6/site-packages/scrapy/spidermiddlewares/depth.py", line 58, in <genexpr>
    return (r for r in result or () if _filter(r))
  File "/home/circleci/zulip/tools/documentation_crawler/documentation_crawler/spiders/common/spiders.py", line 104, in parse
    headers={'Content-Type': response.headers['Content-Type']},
  File "/srv/zulip-venv-cache/1dfe4580e212668e0395ceb9853ebc6f33e66f47/zulip-py3-venv/lib/python3.6/site-packages/scrapy/http/headers.py", line 41, in __getitem__
    return super(Headers, self).__getitem__(key)[-1]
  File "/srv/zulip-venv-cache/1dfe4580e212668e0395ceb9853ebc6f33e66f47/zulip-py3-venv/lib/python3.6/site-packages/scrapy/utils/datatypes.py", line 196, in __getitem__
    return dict.__getitem__(self, self.normkey(key))
KeyError: b'Content-Type'
Passed!
```

so I’m going to investigate this more.

> I think it might be a good idea regardless to move the external links tests

Probably, although I’m pretty sure this still does validation even with `--skip-links` or `--skip-external-links`.