@andersk 2019-06-19
Updated with fixes for three new HTML errors in master since last time. On the other hand, some of the errors I had to ignore last time have been fixed in master since then. I also did some light refactoring to invoke the validator asynchronously through Scrapy rather than synchronously through Requests.

(The error handling problems from my last comment were addressed in #12441.)